@@ -5,15 +5,20 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql;
 
 use Yiisoft\Db\Expression\ArrayExpression;
-use Yiisoft\Db\Expression\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\CaseExpression;
 use Yiisoft\Db\Expression\JsonExpression;
+use Yiisoft\Db\Expression\StructuredExpression;
 use Yiisoft\Db\Pgsql\Builder\ArrayExpressionBuilder;
+use Yiisoft\Db\Pgsql\Builder\ArrayOverlapsConditionBuilder;
+use Yiisoft\Db\Pgsql\Builder\CaseExpressionBuilder;
+use Yiisoft\Db\Pgsql\Builder\JsonOverlapsConditionBuilder;
+use Yiisoft\Db\Pgsql\Builder\LikeConditionBuilder;
 use Yiisoft\Db\Pgsql\Builder\StructuredExpressionBuilder;
 use Yiisoft\Db\Pgsql\Builder\JsonExpressionBuilder;
 use Yiisoft\Db\QueryBuilder\AbstractDQLQueryBuilder;
+use Yiisoft\Db\QueryBuilder\Condition\ArrayOverlapsCondition;
+use Yiisoft\Db\QueryBuilder\Condition\JsonOverlapsCondition;
 use Yiisoft\Db\QueryBuilder\Condition\LikeCondition;
-
-use function array_merge;
 
 /**
  * Implements a DQL (Data Query Language) SQL statements for PostgreSQL Server.
@@ -29,29 +34,26 @@ final class DQLQueryBuilder extends AbstractDQLQueryBuilder
      */
     protected function defaultConditionClasses(): array
     {
-        return array_merge(parent::defaultConditionClasses(), [
+        return [
+            ...parent::defaultConditionClasses(),
             'ILIKE' => LikeCondition::class,
             'NOT ILIKE' => LikeCondition::class,
             'OR ILIKE' => LikeCondition::class,
             'OR NOT ILIKE' => LikeCondition::class,
-        ]);
+        ];
     }
 
-    /**
-     * Has an array of default expression builders.
-     *
-     * Extend this method and override it if you want to change default expression builders for this query builder.
-     *
-     * {@see ExpressionBuilder} docs for details.
-     *
-     * @psalm-return array<string, class-string<ExpressionBuilderInterface>>
-     */
     protected function defaultExpressionBuilders(): array
     {
-        return array_merge(parent::defaultExpressionBuilders(), [
+        return [
+            ...parent::defaultExpressionBuilders(),
             ArrayExpression::class => ArrayExpressionBuilder::class,
+            ArrayOverlapsCondition::class => ArrayOverlapsConditionBuilder::class,
             JsonExpression::class => JsonExpressionBuilder::class,
+            JsonOverlapsCondition::class => JsonOverlapsConditionBuilder::class,
             StructuredExpression::class => StructuredExpressionBuilder::class,
-        ]);
+            LikeCondition::class => LikeConditionBuilder::class,
+            CaseExpression::class => CaseExpressionBuilder::class,
+        ];
     }
 }
