@@ -6,6 +6,8 @@ namespace Yiisoft\Db\Pgsql\Tests\Provider;
 
 use Yiisoft\Db\Expression\ArrayExpression;
 use Yiisoft\Db\Expression\JsonExpression;
+use Yiisoft\Db\Pgsql\Column\ColumnBuilder;
+use Yiisoft\Db\Pgsql\IndexMethod;
 use Yiisoft\Db\Pgsql\Tests\Support\TestTrait;
 
 final class CommandProvider extends \Yiisoft\Db\Tests\Provider\CommandProvider
@@ -18,60 +20,61 @@ final class CommandProvider extends \Yiisoft\Db\Tests\Provider\CommandProvider
     {
         $batchInsert = parent::batchInsert();
 
-        $batchInsert['batchInsert binds params from jsonExpression'] = [
+        $batchInsert['binds json params']['expected'] =
+            'INSERT INTO "type" ("int_col", "char_col", "float_col", "bool_col", "json_col")'
+            . ' VALUES (1, :qp0, 0, TRUE, :qp1::json), (2, :qp2, -1, FALSE, :qp3)';
+
+        $batchInsert['binds params from jsonExpression'] = [
             '{{%type}}',
-            ['json_col', 'int_col', 'float_col', 'char_col', 'bool_col'],
             [
                 [
                     new JsonExpression(
                         ['username' => 'silverfire', 'is_active' => true, 'langs' => ['Ukrainian', 'Russian', 'English']]
                     ),
                     1,
-                    1,
+                    1.0,
                     '',
                     false,
                 ],
             ],
+            ['json_col', 'int_col', 'float_col', 'char_col', 'bool_col'],
             'expected' => <<<SQL
-            INSERT INTO "type" ("json_col", "int_col", "float_col", "char_col", "bool_col") VALUES (:qp0, :qp1, :qp2, :qp3, :qp4)
+            INSERT INTO "type" ("json_col", "int_col", "float_col", "char_col", "bool_col") VALUES (:qp0, 1, 1, :qp1, FALSE)
             SQL,
             'expectedParams' => [
                 ':qp0' => '{"username":"silverfire","is_active":true,"langs":["Ukrainian","Russian","English"]}',
-                ':qp1' => 1,
-                ':qp2' => 1.0,
-                ':qp3' => '',
-                ':qp4' => false,
+                ':qp1' => '',
             ],
         ];
 
-        $batchInsert['batchInsert binds params from arrayExpression'] = [
+        $batchInsert['binds params from arrayExpression'] = [
             '{{%type}}',
+            [[new ArrayExpression([1, null, 3], 'int'), 1, 1.0, '', false]],
             ['intarray_col', 'int_col', 'float_col', 'char_col', 'bool_col'],
-            [[new ArrayExpression([1,null,3], 'int'), 1, 1, '', false]],
             'expected' => <<<SQL
-            INSERT INTO "type" ("intarray_col", "int_col", "float_col", "char_col", "bool_col") VALUES (ARRAY[:qp0, :qp1, :qp2]::int[], :qp3, :qp4, :qp5, :qp6)
+            INSERT INTO "type" ("intarray_col", "int_col", "float_col", "char_col", "bool_col") VALUES (ARRAY[1,NULL,3]::int[], 1, 1, :qp0, FALSE)
             SQL,
-            'expectedParams' => [':qp0' => 1, ':qp1' => null, ':qp2' => 3, ':qp3' => 1, ':qp4' => 1.0, ':qp5' => '', ':qp6' => false],
+            'expectedParams' => [':qp0' => ''],
         ];
 
-        $batchInsert['batchInsert casts string to int according to the table schema'] = [
+        $batchInsert['casts string to int according to the table schema'] = [
             '{{%type}}',
-            ['int_col', 'float_col', 'char_col', 'bool_col'],
             [['3', '1.1', '', false]],
+            ['int_col', 'float_col', 'char_col', 'bool_col'],
             'expected' => <<<SQL
-            INSERT INTO "type" ("int_col", "float_col", "char_col", "bool_col") VALUES (:qp0, :qp1, :qp2, :qp3)
+            INSERT INTO "type" ("int_col", "float_col", "char_col", "bool_col") VALUES (3, 1.1, :qp0, FALSE)
             SQL,
-            'expectedParams' => [':qp0' => 3, ':qp1' => 1.1, ':qp2' => '', ':qp3' => false],
+            'expectedParams' => [':qp0' => ''],
         ];
 
-        $batchInsert['batchInsert binds params from jsonbExpression'] = [
+        $batchInsert['binds params from jsonbExpression'] = [
             '{{%type}}',
-            ['jsonb_col', 'int_col', 'float_col', 'char_col', 'bool_col'],
             [[new JsonExpression(['a' => true]), 1, 1.1, '', false]],
+            ['jsonb_col', 'int_col', 'float_col', 'char_col', 'bool_col'],
             'expected' => <<<SQL
-            INSERT INTO "type" ("jsonb_col", "int_col", "float_col", "char_col", "bool_col") VALUES (:qp0, :qp1, :qp2, :qp3, :qp4)
+            INSERT INTO "type" ("jsonb_col", "int_col", "float_col", "char_col", "bool_col") VALUES (:qp0, 1, 1.1, :qp1, FALSE)
             SQL,
-            'expectedParams' => [':qp0' => '{"a":true}', ':qp1' => 1, ':qp2' => 1.1, ':qp3' => '', ':qp4' => false],
+            'expectedParams' => [':qp0' => '{"a":true}', ':qp1' => ''],
         ];
 
 
@@ -89,5 +92,18 @@ final class CommandProvider extends \Yiisoft\Db\Tests\Provider\CommandProvider
                 SQL,
             ],
         ]);
+    }
+
+    public static function createIndex(): array
+    {
+        return [
+            ...parent::createIndex(),
+            [['col1' => ColumnBuilder::integer()], ['col1'], null, IndexMethod::BTREE],
+            [['col1' => ColumnBuilder::integer()], ['col1'], null, IndexMethod::HASH],
+            [['col1' => ColumnBuilder::integer()], ['col1'], null, IndexMethod::BRIN],
+            [['col1' => ColumnBuilder::array()], ['col1'], null, IndexMethod::GIN],
+            [['col1' => 'point'], ['col1'], null, IndexMethod::GIST],
+            [['col1' => 'point'], ['col1'], null, IndexMethod::SPGIST],
+        ];
     }
 }
