@@ -4,359 +4,196 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql\Tests\Provider;
 
-use ArrayIterator;
-use Yiisoft\Db\Pgsql\Tests\Support\ColumnSchemaBuilder;
-use Yiisoft\Db\Tests\Support\TraversableObject;
+use Yiisoft\Db\Constant\ColumnType;
+use Yiisoft\Db\Pgsql\Column\ArrayColumn;
+use Yiisoft\Db\Pgsql\Column\IntegerColumn;
+use Yiisoft\Db\Pgsql\Column\StructuredColumn;
+use Yiisoft\Db\Schema\Column\DoubleColumn;
+use Yiisoft\Db\Schema\Column\StringColumn;
 
-final class StructuredTypeProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
+final class StructuredTypeProvider
 {
     public static function columns(): array
     {
         return [
             [
                 [
-                    'id' => [
-                        'type' => 'integer',
-                        'dbType' => 'int4',
-                        'phpType' => 'integer',
-                        'primaryKey' => true,
-                        'allowNull' => false,
-                        'autoIncrement' => true,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => 32,
-                        'scale' => 0,
-                        'defaultValue' => null,
-                    ],
-                    'price_col' => [
-                        'type' => 'structured',
-                        'dbType' => 'currency_money_structured',
-                        'phpType' => 'array',
-                        'primaryKey' => false,
-                        'allowNull' => true,
-                        'autoIncrement' => false,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => null,
-                        'scale' => null,
-                        'defaultValue' => null,
-                        'columns' => [
-                            'value' => [
-                                'type' => 'decimal',
-                                'dbType' => 'numeric',
-                                'phpType' => 'double',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => 10,
-                                'scale' => 2,
-                                'defaultValue' => null,
-                            ],
-                            'currency_code' => [
-                                'type' => 'char',
-                                'dbType' => 'bpchar',
-                                'phpType' => 'string',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => 3,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
-                            ],
+                    'id' => new IntegerColumn(
+                        dbType: 'int4',
+                        primaryKey: true,
+                        notNull: true,
+                        autoIncrement: true,
+                        sequenceName: 'test_structured_type_id_seq',
+                        scale: 0,
+                    ),
+                    'price_col' => new StructuredColumn(
+                        dbType: 'currency_money_structured',
+                        defaultValue: null,
+                        columns: [
+                            'value' => new DoubleColumn(
+                                ColumnType::DECIMAL,
+                                dbType: 'numeric',
+                                name: 'value',
+                                notNull: false,
+                                size: 10,
+                                scale: 2,
+                                defaultValue: null,
+                            ),
+                            'currency_code' => new StringColumn(
+                                ColumnType::CHAR,
+                                dbType: 'bpchar',
+                                name: 'currency_code',
+                                notNull: false,
+                                size: 3,
+                                defaultValue: null,
+                            ),
                         ],
-                    ],
-                    'price_default' => [
-                        'type' => 'structured',
-                        'dbType' => 'currency_money_structured',
-                        'phpType' => 'array',
-                        'primaryKey' => false,
-                        'allowNull' => true,
-                        'autoIncrement' => false,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => null,
-                        'scale' => null,
-                        'defaultValue' => ['value' => 5.0, 'currency_code' => 'USD'],
-                        'columns' => [
-                            'value' => [
-                                'type' => 'decimal',
-                                'dbType' => 'numeric',
-                                'phpType' => 'double',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => 10,
-                                'scale' => 2,
-                                'defaultValue' => null,
-                            ],
-                            'currency_code' => [
-                                'type' => 'char',
-                                'dbType' => 'bpchar',
-                                'phpType' => 'string',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => 3,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
-                            ],
+                    ),
+                    'price_default' => new StructuredColumn(
+                        dbType: 'currency_money_structured',
+                        defaultValue: ['value' => 5.0, 'currency_code' => 'USD'],
+                        columns: [
+                            'value' => new DoubleColumn(
+                                ColumnType::DECIMAL,
+                                dbType: 'numeric',
+                                defaultValue: 5.0,
+                                name: 'value',
+                                notNull: false,
+                                size: 10,
+                                scale: 2,
+                            ),
+                            'currency_code' => new StringColumn(
+                                ColumnType::CHAR,
+                                dbType: 'bpchar',
+                                defaultValue: 'USD',
+                                name: 'currency_code',
+                                notNull: false,
+                                size: 3,
+                            ),
                         ],
-                    ],
-                    'price_array' => [
-                        'type' => 'structured',
-                        'dbType' => 'currency_money_structured',
-                        'phpType' => 'array',
-                        'primaryKey' => false,
-                        'allowNull' => true,
-                        'autoIncrement' => false,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => null,
-                        'scale' => null,
-                        'defaultValue' => [
+                    ),
+                    'price_array' => new ArrayColumn(
+                        dbType: 'currency_money_structured',
+                        defaultValue: [
                             null,
                             ['value' => 10.55, 'currency_code' => 'USD'],
                             ['value' => -1.0, 'currency_code' => null],
                         ],
-                        'dimension' => 1,
-                        'columns' => [
-                            'value' => [
-                                'type' => 'decimal',
-                                'dbType' => 'numeric',
-                                'phpType' => 'double',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => 10,
-                                'scale' => 2,
-                                'defaultValue' => null,
+                        dimension: 1,
+                        column: new StructuredColumn(
+                            dbType: 'currency_money_structured',
+                            columns: [
+                                'value' => new DoubleColumn(
+                                    ColumnType::DECIMAL,
+                                    dbType: 'numeric',
+                                    name: 'value',
+                                    notNull: false,
+                                    size: 10,
+                                    scale: 2,
+                                    defaultValue: null,
+                                ),
+                                'currency_code' => new StringColumn(
+                                    ColumnType::CHAR,
+                                    dbType: 'bpchar',
+                                    name: 'currency_code',
+                                    notNull: false,
+                                    size: 3,
+                                    defaultValue: null,
+                                ),
                             ],
-                            'currency_code' => [
-                                'type' => 'char',
-                                'dbType' => 'bpchar',
-                                'phpType' => 'string',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => 3,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
+                            name: 'price_array',
+                            notNull: false,
+                        ),
+                    ),
+                    'price_array2' => new ArrayColumn(
+                        dbType: 'currency_money_structured',
+                        dimension: 2,
+                        column: new StructuredColumn(
+                            dbType: 'currency_money_structured',
+                            columns: [
+                                'value' => new DoubleColumn(
+                                    ColumnType::DECIMAL,
+                                    dbType: 'numeric',
+                                    name: 'value',
+                                    notNull: false,
+                                    size: 10,
+                                    scale: 2,
+                                    defaultValue: null,
+                                ),
+                                'currency_code' => new StringColumn(
+                                    ColumnType::CHAR,
+                                    dbType: 'bpchar',
+                                    name: 'currency_code',
+                                    notNull: false,
+                                    size: 3,
+                                    defaultValue: null,
+                                ),
                             ],
-                        ],
-                    ],
-                    'price_array2' => [
-                        'type' => 'structured',
-                        'dbType' => 'currency_money_structured',
-                        'phpType' => 'array',
-                        'primaryKey' => false,
-                        'allowNull' => true,
-                        'autoIncrement' => false,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => null,
-                        'scale' => null,
-                        'defaultValue' => null,
-                        'dimension' => 2,
-                        'columns' => [
-                            'value' => [
-                                'type' => 'decimal',
-                                'dbType' => 'numeric',
-                                'phpType' => 'double',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => 10,
-                                'scale' => 2,
-                                'defaultValue' => null,
-                            ],
-                            'currency_code' => [
-                                'type' => 'char',
-                                'dbType' => 'bpchar',
-                                'phpType' => 'string',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => 3,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
-                            ],
-                        ],
-                    ],
-                    'range_price_col' => [
-                        'type' => 'structured',
-                        'dbType' => 'range_price_structured',
-                        'phpType' => 'array',
-                        'primaryKey' => false,
-                        'allowNull' => true,
-                        'autoIncrement' => false,
-                        'enumValues' => null,
-                        'size' => null,
-                        'precision' => null,
-                        'scale' => null,
-                        'defaultValue' => [
+                            name: 'price_array2',
+                            notNull: false,
+                        ),
+                    ),
+                    'range_price_col' => new StructuredColumn(
+                        dbType: 'range_price_structured',
+                        defaultValue: [
                             'price_from' => ['value' => 0.0, 'currency_code' => 'USD'],
                             'price_to' => ['value' => 100.0, 'currency_code' => 'USD'],
                         ],
-                        'dimension' => 0,
-                        'columns' => [
-                            'price_from' => [
-                                'type' => 'structured',
-                                'dbType' => 'currency_money_structured',
-                                'phpType' => 'array',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
-                                'columns' => [
-                                    'value' => [
-                                        'type' => 'decimal',
-                                        'dbType' => 'numeric',
-                                        'phpType' => 'double',
-                                        'primaryKey' => false,
-                                        'allowNull' => true,
-                                        'autoIncrement' => false,
-                                        'enumValues' => null,
-                                        'size' => null,
-                                        'precision' => 10,
-                                        'scale' => 2,
-                                        'defaultValue' => null,
-                                    ],
-                                    'currency_code' => [
-                                        'type' => 'char',
-                                        'dbType' => 'bpchar',
-                                        'phpType' => 'string',
-                                        'primaryKey' => false,
-                                        'allowNull' => true,
-                                        'autoIncrement' => false,
-                                        'enumValues' => null,
-                                        'size' => 3,
-                                        'precision' => null,
-                                        'scale' => null,
-                                        'defaultValue' => null,
-                                    ],
+                        columns: [
+                            'price_from' => new StructuredColumn(
+                                dbType: 'currency_money_structured',
+                                defaultValue: ['value' => 0.0, 'currency_code' => 'USD'],
+                                columns: [
+                                    'value' => new DoubleColumn(
+                                        ColumnType::DECIMAL,
+                                        dbType: 'numeric',
+                                        name: 'value',
+                                        notNull: false,
+                                        size: 10,
+                                        scale: 2,
+                                        defaultValue: 0.0,
+                                    ),
+                                    'currency_code' => new StringColumn(
+                                        ColumnType::CHAR,
+                                        dbType: 'bpchar',
+                                        name: 'currency_code',
+                                        notNull: false,
+                                        size: 3,
+                                        defaultValue: 'USD',
+                                    ),
                                 ],
-                            ],
-                            'price_to' => [
-                                'type' => 'structured',
-                                'dbType' => 'currency_money_structured',
-                                'phpType' => 'array',
-                                'primaryKey' => false,
-                                'allowNull' => true,
-                                'autoIncrement' => false,
-                                'enumValues' => null,
-                                'size' => null,
-                                'precision' => null,
-                                'scale' => null,
-                                'defaultValue' => null,
-                                'columns' => [
-                                    'value' => [
-                                        'type' => 'decimal',
-                                        'dbType' => 'numeric',
-                                        'phpType' => 'double',
-                                        'primaryKey' => false,
-                                        'allowNull' => true,
-                                        'autoIncrement' => false,
-                                        'enumValues' => null,
-                                        'size' => null,
-                                        'precision' => 10,
-                                        'scale' => 2,
-                                        'defaultValue' => null,
-                                    ],
-                                    'currency_code' => [
-                                        'type' => 'char',
-                                        'dbType' => 'bpchar',
-                                        'phpType' => 'string',
-                                        'primaryKey' => false,
-                                        'allowNull' => true,
-                                        'autoIncrement' => false,
-                                        'enumValues' => null,
-                                        'size' => 3,
-                                        'precision' => null,
-                                        'scale' => null,
-                                        'defaultValue' => null,
-                                    ],
+                                name: 'price_from',
+                                notNull: false,
+                            ),
+                            'price_to' => new StructuredColumn(
+                                dbType: 'currency_money_structured',
+                                defaultValue: ['value' => 100.0, 'currency_code' => 'USD'],
+                                columns: [
+                                    'value' => new DoubleColumn(
+                                        ColumnType::DECIMAL,
+                                        dbType: 'numeric',
+                                        name: 'value',
+                                        notNull: false,
+                                        size: 10,
+                                        scale: 2,
+                                        defaultValue: 100.0,
+                                    ),
+                                    'currency_code' => new StringColumn(
+                                        ColumnType::CHAR,
+                                        dbType: 'bpchar',
+                                        name: 'currency_code',
+                                        notNull: false,
+                                        size: 3,
+                                        defaultValue: 'USD',
+                                    ),
                                 ],
-                            ],
+                                name: 'price_to',
+                                notNull: false,
+                            ),
                         ],
-                    ],
+                    ),
                 ],
                 'test_structured_type',
-            ],
-        ];
-    }
-
-    public static function normolizedValues()
-    {
-        $price5UsdColumns = [
-            'value' => ColumnSchemaBuilder::numeric(name: 'value', precision: 10, scale: 2, defaultValue: 5.0),
-            'currency_code' => ColumnSchemaBuilder::char(name: 'currency_code', size: 3, defaultValue: 'USD'),
-        ];
-
-        return [
-            'Sort according to `$columns` order' => [
-                ['currency_code' => 'USD', 'value' => 10.0],
-                ['value' => 10.0, 'currency_code' => 'USD'],
-                $price5UsdColumns,
-            ],
-            'Remove excessive elements' => [
-                ['value' => 10.0, 'currency_code' => 'USD', 'excessive' => 'element'],
-                ['value' => 10.0, 'currency_code' => 'USD'],
-                $price5UsdColumns,
-            ],
-            'Fill default values for skipped fields' => [
-                ['currency_code' => 'CNY'],
-                ['value' => 5.0, 'currency_code' => 'CNY'],
-                $price5UsdColumns,
-            ],
-            'Fill default values and column names for skipped indexed fields' => [
-                [10.0],
-                ['value' => 10.0, 'currency_code' => 'USD'],
-                $price5UsdColumns,
-            ],
-            'Fill default values and column names for iterable object' => [
-                new TraversableObject([10.0]),
-                ['value' => 10.0, 'currency_code' => 'USD'],
-                $price5UsdColumns,
-            ],
-            'Fill default values for iterable object' => [
-                new ArrayIterator(['currency_code' => 'CNY']),
-                ['value' => 5.0, 'currency_code' => 'CNY'],
-                $price5UsdColumns,
-            ],
-            'Fill default values for empty array' => [
-                [],
-                ['value' => 5.0, 'currency_code' => 'USD'],
-                $price5UsdColumns,
-            ],
-            'Do not normalize scalar values' => [
-                1,
-                1,
-                $price5UsdColumns,
-            ],
-            'Do not normalize with empty columns' => [
-                [10.0],
-                [10.0],
-                [],
             ],
         ];
     }
