@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql\Tests\Support;
 
-use Yiisoft\Db\Driver\Pdo\PdoConnectionInterface;
-use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Pgsql\Connection;
 use Yiisoft\Db\Pgsql\Driver;
 use Yiisoft\Db\Pgsql\Dsn;
@@ -17,15 +14,18 @@ trait TestTrait
     private string $dsn = '';
     private string $fixture = 'pgsql.sql';
 
-    /**
-     * @throws InvalidConfigException
-     * @throws Exception
-     */
-    protected function getConnection(bool $fixture = false): PdoConnectionInterface
+    public static function setUpBeforeClass(): void
     {
-        $pdoDriver = new Driver($this->getDsn(), 'root', 'root');
-        $pdoDriver->charset('utf8');
-        $db = new Connection($pdoDriver, DbHelper::getSchemaCache());
+        $db = self::getDb();
+
+        DbHelper::loadFixture($db, __DIR__ . '/Fixture/pgsql.sql');
+
+        $db->close();
+    }
+
+    protected function getConnection(bool $fixture = false): Connection
+    {
+        $db = new Connection($this->getDriver(), DbHelper::getSchemaCache());
 
         if ($fixture) {
             DbHelper::loadFixture($db, __DIR__ . "/Fixture/$this->fixture");
@@ -34,23 +34,33 @@ trait TestTrait
         return $db;
     }
 
-    protected static function getDb(): PdoConnectionInterface
+    protected static function getDb(): Connection
     {
-        $dsn = (new Dsn('pgsql', '127.0.0.1', 'yiitest', '5432'))->asString();
+        $dsn = (new Dsn(
+            host: self::getHost(),
+            databaseName: self::getDatabaseName(),
+            port: self::getPort(),
+        ))->asString();
+        $driver = new Driver($dsn, self::getUsername(), self::getPassword());
+        $driver->charset('utf8');
 
-        return new Connection(new Driver($dsn, 'root', 'root'), DbHelper::getSchemaCache());
+        return new Connection($driver, DbHelper::getSchemaCache());
     }
 
     protected function getDsn(): string
     {
         if ($this->dsn === '') {
-            $this->dsn = (new Dsn('pgsql', '127.0.0.1', 'yiitest', '5432'))->asString();
+            $this->dsn = (new Dsn(
+                host: self::getHost(),
+                databaseName: self::getDatabaseName(),
+                port: self::getPort(),
+            ))->asString();
         }
 
         return $this->dsn;
     }
 
-    protected function getDriverName(): string
+    protected static function getDriverName(): string
     {
         return 'pgsql';
     }
@@ -63,5 +73,38 @@ trait TestTrait
     protected function setFixture(string $fixture): void
     {
         $this->fixture = $fixture;
+    }
+
+    protected function getDriver(): Driver
+    {
+        $driver = new Driver($this->getDsn(), self::getUsername(), self::getPassword());
+        $driver->charset('utf8');
+
+        return $driver;
+    }
+
+    private static function getDatabaseName(): string
+    {
+        return getenv('YII_PGSQL_DATABASE') ?: 'yiitest';
+    }
+
+    private static function getHost(): string
+    {
+        return getenv('YII_PGSQL_HOST') ?: '127.0.0.1';
+    }
+
+    private static function getPort(): string
+    {
+        return getenv('YII_PGSQL_PORT') ?: '5432';
+    }
+
+    private static function getUsername(): string
+    {
+        return getenv('YII_PGSQL_USER') ?: 'root';
+    }
+
+    private static function getPassword(): string
+    {
+        return getenv('YII_PGSQL_PASSWORD') ?: 'root';
     }
 }
